add StateMachine interface
1. add StateMachine interface
2. add a local disk based implementaion

StateMachine represents a collection of commited logs. Once a log
entry is commited into StateMachine, it means this log will be appied
to all servers eventually, in the same order. In other words, once a
log entry is commited to StateMachine, it is safely to say the cluster
make an agreement on this log and client's request is fulfilled.