add TimerService
A TimerService basically runs a user specified code for timeout event.
Typically, user passes a Runnable to a TimerService and optionally
a timeout, so that TimerService will run the code once timeout event
happens. In addition, TimerService provide a way for cancelation, if
user call reset before TimerService running the code, then TimerService
should reschedule the execution and wait for another timeout event.