Add Communicator
A Communicator provides functionalities to communicate with other
peers. In YRaft, each server needs to send/receive message from/to
other peers, a communicator instance will be passed to RaftServer so
that it does not need to worry about the internal details to deal
with message sending/receiving.

Currently, a actor based communicator is implementated.